Move blockHandlers.js to lib/utilities
