+++ src/app/containers/Image/index.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { filterForBlockType } from '../../lib/blockHandlers';
+import filterForBlockType from '../../lib/utilities/blockHandlers';
 import { imageModelPropTypes } from '../../models/propTypes/image';
 import Figure from '../Figure';
 import {
