+++ src/app/lib/utilities/blockHandlers.js
@@ -3,8 +3,7 @@
 // elements and blocks within them
 
 // Filters array of blocks for a single block of given type
-export const filterForBlockType = (arrayOfBlocks, type) =>
+const filterForBlockType = (arrayOfBlocks, type) =>
   arrayOfBlocks.filter(block => (block ? block.type === type : null))[0];
 
-export const extractText = arrayOfBlocks =>
-  arrayOfBlocks[0].model.blocks[0].model;
+export default filterForBlockType;
